+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 
 ### Bug fixes
 
-
+- The skeleton projection layer will now update automatically on skeleton
+  changes like new or removed nodes as well as splits and merges.
 
 
 ## 2015.10.19